refactor(ka): split alignment_test.go into per-concern files (QE-2)
Splits the 2.1k-line monolithic test file into 13 focused files:
- helpers_test.go: shared mocks and helper functions
- 9 per-concern test files organized by feature area
- Original alignment_test.go reduced to 340 lines (evaluator core)

All 107 specs preserved and passing with race detector.
